backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
feedparser==6.0.11
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
pydantic==2.10.3
pydantic_core==2.27.1
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
cryptography==44.0.0
redis==5.2.1
loguru==0.7.0
prometheus_client==0.17.1
argon2-cffi==23.1.0
pytest==7.3.1
httpx==0.24.1
black==23.3.0
